requirements.txt
numpy==1.19.3
matplotlib==3.3.2
python-dateutil==2.8.1
rpy2==3.3.6
netCDF4==1.5.4
scipy==1.5.3
Cartopy==0.18.0
arm-pyart==1.11.2
wradlib==1.8.1
